deps/wazuh_testing/wazuh_testing/qa_ctl/deployment/windows/requirements.txt
requests
lockfile
setuptools
python-vagrant
pyyaml
